functions/check-gcp-existence/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
cachetools==4.2.2
    # via google-auth
certifi==2021.5.30
    # via requests
cffi==1.14.6
    # via
    #   cryptography
    #   google-crc32c
chardet==4.0.0
    # via requests
ckanapi==4.5
    # via -r requirements.in
cryptography==3.4.7
    # via oauthlib
dataclasses==0.6
    # via -r requirements.in
decorator==5.0.9
    # via retry
defusedxml==0.7.1
    # via jira
docopt==0.6.2
    # via ckanapi
gobits==1.0.8
    # via -r requirements.in
google-api-core[grpc]==1.31.0
    # via
    #   google-api-python-client
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-logging
    #   google-cloud-pubsub
    #   google-cloud-secret-manager
google-api-python-client==1.12.5
    # via -r requirements.in
google-auth==1.33.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-cloud-storage
google-auth-httplib2==0.1.0
    # via google-api-python-client
google-cloud-bigquery==2.11.0
    # via -r requirements.in
google-cloud-core==1.4.3
    # via
    #   -r requirements.in
    #   google-cloud-bigquery
    #   google-cloud-logging
    #   google-cloud-resource-manager
    #   google-cloud-storage
google-cloud-logging==1.15.1
    # via -r requirements.in
google-cloud-pubsub==2.3.0
    # via -r requirements.in
google-cloud-resource-manager==0.30.3
    # via -r requirements.in
google-cloud-secret-manager==2.0.0
    # via -r requirements.in
google-cloud-storage==1.32.0
    # via -r requirements.in
google-crc32c==1.1.2
    # via google-resumable-media
google-resumable-media==1.3.1
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
googleapis-common-protos[grpc]==1.53.0
    # via
    #   google-api-core
    #   grpc-google-iam-v1
grpc-google-iam-v1==0.12.3
    # via
    #   google-cloud-pubsub
    #   google-cloud-secret-manager
grpcio==1.38.1
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
httplib2==0.19.1
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==2.10
    # via requests
jira==2.0.0
    # via -r requirements.in
libcst==0.3.19
    # via
    #   google-cloud-pubsub
    #   google-cloud-secret-manager
mypy-extensions==0.4.3
    # via typing-inspect
oauthlib[signedtoken]==3.1.1
    # via
    #   jira
    #   requests-oauthlib
packaging==21.0
    # via
    #   google-api-core
    #   google-cloud-bigquery
pbr==5.6.0
    # via jira
proto-plus==1.19.0
    # via
    #   google-cloud-bigquery
    #   google-cloud-pubsub
    #   google-cloud-secret-manager
protobuf==3.17.3
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   googleapis-common-protos
    #   proto-plus
py==1.10.0
    # via retry
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pycparser==2.20
    # via cffi
pyjwt==2.1.0
    # via oauthlib
pyparsing==2.4.7
    # via
    #   httplib2
    #   packaging
python-slugify==5.0.2
    # via ckanapi
pytz==2021.1
    # via google-api-core
pyyaml==5.4.1
    # via libcst
requests==2.25.1
    # via
    #   -r requirements.in
    #   ckanapi
    #   google-api-core
    #   google-cloud-storage
    #   jira
    #   requests-oauthlib
    #   requests-toolbelt
requests-oauthlib==1.3.0
    # via jira
requests-toolbelt==0.9.1
    # via jira
retry==0.9.2
    # via -r requirements.in
rsa==4.7.2
    # via google-auth
six==1.16.0
    # via
    #   ckanapi
    #   google-api-core
    #   google-api-python-client
    #   google-auth
    #   google-auth-httplib2
    #   google-resumable-media
    #   grpcio
    #   jira
    #   protobuf
text-unidecode==1.3
    # via python-slugify
typing-extensions==3.10.0.0
    # via
    #   libcst
    #   typing-inspect
typing-inspect==0.7.1
    # via libcst
uritemplate==3.0.1
    # via google-api-python-client
urllib3==1.26.6
    # via requests
werkzeug==1.0.1
    # via gobits

# The following packages are considered to be unsafe in a requirements file:
# setuptools
